requirements.txt
streamlit
Pillow
openai
matplotlib
pandas
numpy
tqdm
seaborn
regex
konlpy
WordCloud
